log rebuildData exceptions; rename setup and override
