+++ config/template/preLocalSettings_allWikis.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * setup.php is used to modify all wiki settings. This is where any
+ * preLocalSettings.php is used to modify all wiki settings. This is where any
  * any changes to the standard meza configuration should be kept. It also
  * is where database user and password info is stored
  **/
