+++ config/template/wiki-init/config/postLocalSettings.php
