+++ config/template/wiki-init/config/preLocalSettings.php
