+++ scripts/elastic-build-index.sh
@@ -2,15 +2,15 @@ bash printTitle.sh "Begin $0"
 
 echo "******* Generating elasticsearch index *******"
 
-# disable search update in wiki-specific overrides
-echo "\$wgDisableSearchUpdate = true;" >> "$m_htdocs/wikis/$wiki_id/config/overrides.php"
+# disable search update in wiki-specific postLocalSettings
+echo "\$wgDisableSearchUpdate = true;" >> "$m_htdocs/wikis/$wiki_id/config/postLocalSettings.php"
 
 # Run script to generate elasticsearch index
 cd "$m_mediawiki"
 WIKI="$wiki_id" php "$m_mediawiki/extensions/CirrusSearch/maintenance/updateSearchIndexConfig.php" --startOver
 
-# Remove search-update disable in wiki-specific overrides
-sed -r -i 's/\$wgDisableSearchUpdate = true;//g;' "$m_htdocs/wikis/$wiki_id/config/overrides.php"
+# Remove search-update disable in wiki-specific postLocalSettings
+sed -r -i 's/\$wgDisableSearchUpdate = true;//g;' "$m_htdocs/wikis/$wiki_id/config/postLocalSettings.php"
 
 # Bootstrap the search index
 #
